CHARACTER CREATOR GUI BETA
This is the big update you've all been waiting for!

This finally adds the GUI I've been working on for months.
You can now finally see your characters, Create new ones, or delete them.

And set them to Auto-Load, or don't. And you'll be presented with the Character selector each time you log in to the server.

Please keep in mind, this is still in a beta stage. So things might still break.

But I have tested it in my own game, and everything appears to work for me.

Please make sure you report any issues!

Thank you, and enjoy!